+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.mobile",
+    package = "com.infolimo.agenda_dynamique",
 )
 
 android_resource(
     name = "res",
-    package = "com.mobile",
+    package = "com.infolimo.agenda_dynamique",
     res = "src/main/res",
 )
