+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "mobile",
-  "displayName": "mobile"
+  "name": "AgendaDynamique",
+  "displayName": "Agenda Dynamique"
 }
