+++ ios/AgendaDynamique.xcodeproj/xcshareddata/xcschemes/AgendaDynamique.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "mobile.app"
-               BlueprintName = "mobile"
-               ReferencedContainer = "container:mobile.xcodeproj">
+               BuildableName = "AgendaDynamique.app"
+               BlueprintName = "AgendaDynamique"
+               ReferencedContainer = "container:AgendaDynamique.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "mobileTests.xctest"
-               BlueprintName = "mobileTests"
-               ReferencedContainer = "container:mobile.xcodeproj">
+               BuildableName = "AgendaDynamiqueTests.xctest"
+               BlueprintName = "AgendaDynamiqueTests"
+               ReferencedContainer = "container:AgendaDynamique.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "mobile.app"
-            BlueprintName = "mobile"
-            ReferencedContainer = "container:mobile.xcodeproj">
+            BuildableName = "AgendaDynamique.app"
+            BlueprintName = "AgendaDynamique"
+            ReferencedContainer = "container:AgendaDynamique.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "mobile.app"
-            BlueprintName = "mobile"
-            ReferencedContainer = "container:mobile.xcodeproj">
+            BuildableName = "AgendaDynamique.app"
+            BlueprintName = "AgendaDynamique"
+            ReferencedContainer = "container:AgendaDynamique.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
