+++ ios/AgendaDynamiqueTests/AgendaDynamiqueTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface mobileTests : XCTestCase
+@interface AgendaDynamiqueTests : XCTestCase
 
 @end
 
-@implementation mobileTests
+@implementation AgendaDynamiqueTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
