+++ ios/Podfile
@@ -3,7 +3,7 @@ require_relative '../node_modules/@react-native-community/cli-platform-ios/nativ
 
 platform :ios, '10.0'
 
-target 'mobile' do
+target 'AgendaDynamique' do
   config = use_native_modules!
 
   use_react_native!(
@@ -12,7 +12,7 @@ target 'mobile' do
     :hermes_enabled => false
   )
 
-  target 'mobileTests' do
+  target 'AgendaDynamiqueTests' do
     inherit! :complete
     # Pods for testing
   end
