+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "mobile",
+  "name": "AgendaDynamique",
   "version": "0.0.1",
   "private": true,
   "scripts": {
